Added MAC address to host adder dialog
I'm not real happy with how I implemented this. The position of the new field in the entity spec is hard-coded, meaning if the entity spec changes for whatever reason, this code will produce unexpected (and probably unwelcome) results. But I haven't been able to come up with a better way of doing this yet, and right now doing SOMETHING is better than doing it perfectly.